365. Arcane Shards Cost and Cooldown

+++ Source/Aura/Private/AbilitySystem/Abilities/ArcaneShards.cpp
@@ -5,7 +5,7 @@
 
 FString UArcaneShards::GetDescription(int32 Level)
 {
-	const float ScaledDamage = Damage.GetValueAtLevel(Level);
+	const int32 ScaledDamage = Damage.GetValueAtLevel(Level);
 	const float ManaCost = FMath::Abs(GetManaCost(Level));
 	const float Cooldown = GetCooldown(Level);
 
@@ -52,7 +52,7 @@ FString UArcaneShards::GetDescription(int32 Level)
 FString UArcaneShards::GetNextLevelDescription(int32 Level)
 {
 	Level += 1;//next level
-	const float ScaledDamage = Damage.GetValueAtLevel(Level);
+	const int32 ScaledDamage = Damage.GetValueAtLevel(Level);
 	const float ManaCost = FMath::Abs(GetManaCost(Level));
 	const float Cooldown = GetCooldown(Level);
